refactor: moved API model layer moved from database folder to src/models
