+++ src/database/seeders/index.ts
@@ -1,8 +1,8 @@
 /* eslint-disable no-console */
-import AssetModel from '../models/AssetModel';
-import CompanyModel from '../models/CompanyModel';
-import UnitModel from '../models/UnitModel';
-import UserModel from '../models/UserModel';
+import AssetModel from '../../models/AssetModel';
+import CompanyModel from '../../models/CompanyModel';
+import UnitModel from '../../models/UnitModel';
+import UserModel from '../../models/UserModel';
 import assets from './Assets';
 import companies from './Companies';
 import units from './Units';
