+++ src/models/AssetModel.ts
@@ -1,6 +1,6 @@
-import { Asset } from '../../@types/Entities';
-import { IAssetModel } from '../../interfaces/IAssetModel';
-import assetSchema from '../schemas/assetSchema';
+import { Asset } from '../@types/Entities';
+import { IAssetModel } from '../interfaces/IAssetModel';
+import assetSchema from '../database/schemas/assetSchema';
 import Model from './Model';
 
 class AssetModel extends Model<Asset> implements IAssetModel<Asset> {
